MonaBase: - QualityOfService.lastSendingTime property addition - Add String::IsTrue and String::IsFalse and remove String::ToBoolean to get a strict conversion from string to boolean
MonaCore:
- Writer::flush returns a boolean now to indicate if something has been sent
- Fix a duplicate NetStream.Play.PublishNotify event issue
- Merge Thom changes (RTMP disconnection fix + HTTP _index issue)
- Better understandinf of FlashStream::messageHandler and FlashStream::dataHandler: AMF_INVOCATION(_AMF3) => messageHandler, AMF_DATA(_AMF3) => dataHandler
- Better understanding of FlashStream::rawMessage handler, fix a NetGroup::close possible bug (no peer.unjoin event)
- Reorder arguments of onFileAccess method to respect code convention

+++ MonaCore/include/Mona/FlashStream.h
@@ -59,7 +59,7 @@ protected:
 private:
 
 	virtual void	messageHandler(const std::string& name, AMFReader& message, FlashWriter& writer);
-	virtual void	rawHandler(UInt8 type, PacketReader& data, FlashWriter& writer);
+	virtual void	rawHandler(UInt16 type, PacketReader& data, FlashWriter& writer);
 	virtual void	dataHandler(DataReader& data, double lostRate);
 	virtual void	audioHandler(UInt32 time, PacketReader& packet, double lostRate);
 	virtual void	videoHandler(UInt32 time,PacketReader& packet, double lostRate);
